+++ mod-source-record-manager-server/src/main/java/org/folio/services/migration/impl/CancelledLccnMappingRenamingMigration.java
@@ -1,31 +1,17 @@
 package org.folio.services.migration.impl;
 
-import io.vertx.core.Future;
-import org.folio.Record.RecordType;
+import io.vertx.core.json.JsonObject;
 import org.folio.services.MappingRuleService;
-import org.folio.services.migration.CustomMigration;
 
-public abstract class CancelledLccnMappingRenamingMigration implements CustomMigration {
-
-  private final MappingRuleService mappingRuleService;
+public abstract class CancelledLccnMappingRenamingMigration extends BaseMigration {
 
   protected CancelledLccnMappingRenamingMigration(MappingRuleService mappingRuleService) {
-    this.mappingRuleService = mappingRuleService;
+    super(mappingRuleService);
   }
 
   @Override
-  public Future<Void> migrate(String tenantId) {
-    return mappingRuleService.get(getRecordType(), tenantId)
-      .compose(rules -> {
-        if (rules.isPresent()) {
-          var newRules = rules.get().encode().replace("Cancelled LCCN", "Canceled LCCN");
-          return mappingRuleService.internalUpdate(newRules, getRecordType(), tenantId);
-        } else {
-          return Future.succeededFuture();
-        }
-      }).mapEmpty();
+  protected String updateRules(JsonObject rules) {
+    return rules.encode().replace("Cancelled LCCN", "Canceled LCCN");
   }
-
-  protected abstract RecordType getRecordType();
 
 }
